Added modifications and changes to enable fp16/bf16 based compilation and inference. Almost all LLMs can now be compiled and infered in fp16, test_causal_lm_models script has the following notion regarding how the tests happened : # means the model wasn't tested due to the size, not sure if it'll run through or have an accuracy mismatch. ## means the ouputs match for fp16 and things worked fine. ### means, outputs come but don't match properly with HF tokens. #### means they're quantized model and additional effort is needed to enable these. These commits cover almost all LLMs currently supported.

+++ QEfficient/transformers/models/gemma2/modeling_gemma2.py
@@ -447,7 +447,7 @@ class QEffGemma2ForCausalLM(Gemma2ForCausalLM, GenerationMixin):
             logits = logits / self.config.final_logit_softcapping
             logits = torch.tanh(logits)
             logits = logits * self.config.final_logit_softcapping
-
+        logits = logits.float()
         return CausalLMOutputWithPast(
             loss=None,
             logits=logits,
